requirements.txt
surya-ocr
python-dotenv
Pillow
PyMuPDF
numpy
scikit-learn
transformers
torch